% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moodys_single_series.R
\name{convert_moodys}
\alias{convert_moodys}
\title{Convert a Moody's Data Buffet query to a data frame}
\usage{
convert_moodys(resp)
}
\arguments{
\item{resp}{An \code{httr2_response}, or a list of responses, from
\code{\link[=get_moodys_series]{get_moodys_series()}}.}
}
\value{
A tibble, one row per observation.
}
\description{
Convert a Moody's Data Buffet query to a data frame
}
\examples{
\dontrun{
jobs_forecast_202309 <- get_moodys_series("fet.iusa", vintage = "202309") |>
  convert_moodys()
}
}
