% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_moodys_series.R
\name{convert_moodys_search}
\alias{convert_moodys_search}
\title{Convert Moody's Search Results to Data Frame}
\usage{
convert_moodys_search(search_response)
}
\arguments{
\item{search_response}{Object from search_moodys_series()}
}
\value{
Returns a dataframe with search results
}
\description{
Convert Moody's Search Results to Data Frame
}
\examples{
unemployment_search <- search_moodys_series("unemployment rate")
unemployment_df <- convert_moodys_search(unemployment_search)
}
