% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moodys_single_series.R
\name{get_moodys_series}
\alias{get_moodys_series}
\title{Get Moody's Data Buffet series}
\usage{
get_moodys_series(
  mnemonics,
  freq = "0",
  trans = "0",
  vintage = NULL,
  accKey = moodys_key("acc"),
  encKey = moodys_key("enc")
)
}
\arguments{
\item{mnemonics}{A single mnemonic or a character vector of mnemonics. More
than 25 are requested in batches.}

\item{freq}{Frequency conversion code. See \code{\link[=get_moodys_codes]{get_moodys_codes()}}.}

\item{trans}{Transformation code. See the API user guide.}

\item{vintage}{Vintage to request. See \code{\link[=get_moodys_vintages]{get_moodys_vintages()}}.}

\item{accKey}{Moody's API access key. Defaults to the stored key, see
\code{\link[=moodys_key]{moodys_key()}}.}

\item{encKey}{Moody's API encryption key. Defaults to the stored key.}
}
\value{
An \code{httr2_response} for 25 or fewer mnemonics, or a list of
responses when batching. Pass either to \code{\link[=convert_moodys]{convert_moodys()}}.
}
\description{
Retrieves Moody's Data Buffet series by mnemonic. Requests are automatically
split into chunks of 25, the API's per-call limit. Use \code{\link[=convert_moodys]{convert_moodys()}} to
turn the result into a data frame.

Access and encryption keys can be stored with \code{\link[=set_moodys_api_key]{set_moodys_api_key()}} or
passed directly.

API documentation: \url{https://www.economy.com/products/tools/api#db}

Based on Moody's sample code:
\url{https://github.com/moodysanalytics/databuffet-api-codesamples/blob/master/R/Single-Series.R}
}
\examples{
\dontrun{
jobs_forecast_202309 <- get_moodys_series("fet.iusa", vintage = "202309") |>
  convert_moodys()
}
}
