% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{leifa_long}
\alias{leifa_long}
\title{Leifa Jobs}
\format{
A data frame with 396,396 rows and 7 variables:
\describe{
\item{bea_fips}{County FIPS code, 5 digits, on the BEA basis}
\item{county_fips}{County FIPS code, 3 digits}
\item{area}{Area name}
\item{gcode}{SDO industry code}
\item{leifa_group}{Type of worker: \code{"WS"} (wage and salary),
\code{"Prop"} (sole proprietor), or \code{"Jobs"} (all workers, WS + Prop)}
\item{year}{Year the estimate applies to}
\item{total}{Estimated value}
}
}
\source{
Colorado State Demography Office
}
\usage{
leifa_long
}
\description{
Historical Leifa jobs data produced by the Colorado State Demography
Office, in long format.
}
\keyword{datasets}
