% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoecon-package.R
\docType{package}
\name{sdoecon-package}
\alias{sdoecon}
\alias{sdoecon-package}
\title{sdoecon: Colorado State Demography Office Economic Data Tools}
\description{
Tools used by the Colorado State Demography Office to assemble economic and jobs data. Provides wrappers for the Moody's Analytics Data Buffet API, a downloader for the Bureau of Labor Statistics Quarterly Census of Employment and Wages (QCEW), and accessors for the office's economic crosswalks. Crosswalks are served from the SDO Postgres database via the 'sdotools' package rather than bundled with this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gtottenSDO/sdoecon}
  \item Report bugs at \url{https://github.com/gtottenSDO/sdoecon/issues}
}

}
\author{
\strong{Maintainer}: Greg Totten \email{greg.totten@state.co.us}

Authors:
\itemize{
  \item Greg Totten \email{greg.totten@state.co.us}
}

}
\keyword{internal}
