% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{temporary_env}
\alias{temporary_env}
\title{Create a temporary environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
temporary_env
}
\value{
An environment
}
\description{
Create a temporary environment
}
\keyword{internal}
