% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_api_moodys.R
\name{call_api_moodys}
\alias{call_api_moodys}
\title{Call Moody's API}
\usage{
call_api_moodys(
  apiCommand,
  accKey = Sys.getenv("MOODYS_ACC_KEY"),
  encKey = Sys.getenv("MOODYS_ENC_KEY"),
  type = "GET"
)
}
\arguments{
\item{apiCommand}{Moody's API command}

\item{accKey}{Moody's API acccess key (defaults to environment variable if installed using set_moodys_api_key)}

\item{encKey}{Moody's API encryption key (defaults to environment variable if installed using set_moodys_api_key)}

\item{type}{GET or POST command to pass to API}
}
\value{
httr request object
}
\description{
Function to make a call based on the Moody's economy.com API.

Based on the Moody's API documentation:
https://www.economy.com/support/apis/getting-started
}
