% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodys_authenticate.R
\name{check_moodys_token}
\alias{check_moodys_token}
\title{Check Moody's Token}
\usage{
check_moodys_token(
  accKey = Sys.getenv("MOODYS_ACC_KEY"),
  encKey = Sys.getenv("MOODYS_ENC_KEY")
)
}
\arguments{
\item{encKey}{}
}
\description{
Check Moody's Token
}
