% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moodys_single_series.R
\name{convert_moodys}
\alias{convert_moodys}
\title{Convert Moody's Data Buffet series query to an object}
\usage{
convert_moodys(series)
}
\arguments{
\item{series}{Object from get_moodys_series()}

\item{series_type}{Whether the series is a single or multi series query}
}
\value{
Returns a dataframe
}
\description{
Convert Moody's Data Buffet series query to an object
}
\examples{
jobs_forecast_vintage_202309 <- get_moodys_series("fet.iusa",
vintage = "202309") \%>\%
 convert_moodys(type = "xts")
}
