% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalks.R
\docType{data}
\name{denver_boulder_geography_xwalk}
\alias{denver_boulder_geography_xwalk}
\title{Denver Boulder Geography Crosswalk}
\format{
A data frame with 7 rows and 5 variables:
\describe{
\item{county_fips}{County FIPS code}
\item{county_name}{County name}
\item{area}{Label "Denver-Boulder Region" for labeling in other datasets}
\item{area_fips}{FIPS code for the Denver-Boulder Region (500)}
\item{fips_long}{County FIPS codes converted to 5 digits}
}
}
\source{
Colorado State Demography Office Crosswalk
}
\usage{
denver_boulder_geography_xwalk
}
\description{
A crosswalk file containing the counties in the Denver Boulder MSA
}
\keyword{datasets}
