% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalks.R
\docType{data}
\name{gcode_n2_crosswalk}
\alias{gcode_n2_crosswalk}
\title{Gcode to Naics 2 Digit Crosswalk}
\format{
A data frame with 23 rows and 5 variables:
\describe{
\item{ownership_code}{Ownership code}
\item{gcode}{SDO industry code}
\item{gcode_label}{SDO industry label}
\item{naics_code}{NAICS 2 digit code}
\item{naics_label}{NAICS 2 digit label}
}
}
\source{
Colorado State Demography Office Crosswalk
}
\usage{
gcode_n2_crosswalk
}
\description{
A crosswalk file containing the SDO industry codes and the NAICS 2 digit codes
}
\keyword{datasets}
