% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalks.R
\docType{data}
\name{geography_xwalk}
\alias{geography_xwalk}
\title{Geography Crosswalk}
\format{
A data frame with 65 rows and 8 variables:
\describe{
\item{fips_long}{County FIPS codes converted to 5 digits}
\item{county_fips}{County FIPS code}
\item{county_name}{County name}
\item{pmreg_code}{Codes for identifying planning and management regions}
\item{pmreg_name}{Names for identifying planning and management regions}
\item{super_pm_name}{Names for identifying super planning and management regions}
\item{touism_region_code}{Codes for identifying tourism regions}
\item{base_region_code}{Codes for identifying base regions}
}
}
\source{
Colorado State Demography Office Crosswalk
}
\usage{
geography_xwalk
}
\description{
A crosswalk file containing the counties in Colorado for mapping them to regions
}
\keyword{datasets}
