% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bls_qcew.R
\name{get_bls_qcew}
\alias{get_bls_qcew}
\title{Function to get BLS data from the BLS
Quarterly Census of Employment and Wages (QCEW)}
\usage{
get_bls_qcew(
  year = 2022,
  freq = "a",
  call_type = "zip",
  state = "08",
  county = NULL
)
}
\arguments{
\item{year}{Year to download}

\item{freq}{Frequency of data to download, must be either "q" (quarterly) or "a" (annual) when downloading zip files, or be the quarter number (1-4) or "a" (annual) when using the api call}

\item{call_type}{Either "zip" or "api"}

\item{state}{State FIPS code for calling API or filtering downloaded
zip files}

\item{county}{County FIPS code for calling API or filtering downloaded
zip files}
}
\value{
A data frame with the requested data
}
\description{
Description of the data:
https://www.bls.gov/cew/data-overview.htm
Description of downloadable zip files:
https://www.bls.gov/cew/downloadable-data-files.htm
}
\details{
Description of api:
https://www.bls.gov/cew/additional-resources/open-data/home.htm
}
\examples{
qcew_2020_all <- get_bls_qcew(year = 2020, freq = "q", call_type = "zip")
}
