% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basket_moodys.R
\name{get_moodys_basket}
\alias{get_moodys_basket}
\title{Get Moody's basket data}
\usage{
get_moodys_basket(
  BASKET_NAME,
  accKey = Sys.getenv("MOODYS_ACC_KEY"),
  encKey = Sys.getenv("MOODYS_ENC_KEY")
)
}
\arguments{
\item{BASKET_NAME}{Name of basket to download}

\item{accKey}{Moody's API acccess key (defaults to environment variable if installed using set_moodys_api_key)}

\item{encKey}{Moody's API encryption key (defaults to environment variable if installed using set_moodys_api_key)}
}
\value{
Data frame with long format data
}
\description{
Function to get Moody's data for a specific basket.
Data is converted to long format and statewide and Denver Boulder MSA are aggregated.

Function based on Moody's API documentation:
https://www.economy.com/support/apis/getting-started
}
\examples{
moodys_colorado_forecast <- basket_call(BASKET_NAME = "colorado_sector_emp")
}
