% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodys_api_helpers.R
\name{get_moodys_codes}
\alias{get_moodys_codes}
\title{Table of API Codes}
\usage{
get_moodys_codes(
  codetype = c("filetypes", "frequencies"),
  accKey = Sys.getenv("MOODYS_ACC_KEY"),
  encKey = Sys.getenv("MOODYS_ENC_KEY")
)
}
\arguments{
\item{codetype}{Either "filetypes" or "frequencies" (filetypes are only for baskets)}

\item{accKey}{Access Key for Moody's API (defaults to environment variable)}

\item{encKey}{Encryption Key for Moody's API (defaults to environment variable)}
}
\value{
Data frame of available codes
}
\description{
Table of API Codes
}
