% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moodys_single_series.R
\name{get_moodys_series}
\alias{get_moodys_series}
\title{Get Moody's Data Buffet Series}
\usage{
get_moodys_series(
  mnemonics,
  freq = "0",
  trans = "0",
  vintage = NULL,
  accKey = Sys.getenv("MOODYS_ACC_KEY"),
  encKey = Sys.getenv("MOODYS_ENC_KEY")
)
}
\arguments{
\item{mnemonics}{Either a single mnemonic or list of mnemonics (less than 25)
to pass to the Moody's API}

\item{freq}{Frequency conversion (for list of codes use
get_moodys_codes(frequencies), or see the API user guide)}

\item{trans}{Transformation code (for list of codes see API user guide)}

\item{vintage}{Get list of available vintages using
get_moodys_vintages(mnemonic)}

\item{accKey}{Access Key for Moody's API (defaults to environment variable)}

\item{encKey}{Encryption Key for Moody's API (defaults to environment variable)}
}
\value{
returns list from json. Use function convert_moodys to create
a data frame
}
\description{
This function retrieves Moody's Data Buffet Series using Mnemonic Code
(or set of up to 25 Mnemonic Codes). Access and Encryption Keys can be stored
using set_moodys_api_key(), or passed through to the function directly.
Frequency, Transformation Code, and Vintage arguments are optional. Use
convert_moodys() to create an xts or data frame object.
}
\details{
API documentation can be found at https://www.economy.com/products/tools/api#db

Script based off of Moody's sample code available at:
https://github.com/moodysanalytics/databuffet-api-codesamples/blob/master/R/Single-Series.R
}
\examples{
jobs_forecast_vintage_202309 <- get_moodys_series("fet.iusa",
vintage = "202309") \%>\%
 convert_moodys(type = "xts")
}
