% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodys_authenticate.R
\name{get_moodys_token}
\alias{get_moodys_token}
\title{Authenticate Moody's API for token}
\usage{
get_moodys_token(
  accKey = Sys.getenv("MOODYS_ACC_KEY"),
  encKey = Sys.getenv("MOODYS_ENC_KEY")
)
}
\value{
Stores token and token expiration time in temporary environment variable
}
\description{
Stores token and token expiration time in temporary environment variable
}
