% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodys_api_helpers.R
\name{get_moodys_vintages}
\alias{get_moodys_vintages}
\title{Lookup Available Vintages for a Moodys Mnemonic}
\usage{
get_moodys_vintages(
  mnemonic,
  accKey = Sys.getenv("MOODYS_ACC_KEY"),
  encKey = Sys.getenv("MOODYS_ENC_KEY")
)
}
\arguments{
\item{mnemonic}{Mnemonic to lookup}

\item{accKey}{Access Key for Moody's API (defaults to environment variable)}

\item{encKey}{Encryption Key for Moody's API (defaults to environment variable)}
}
\value{
Data frame of available vintages
}
\description{
Lookup Available Vintages for a Moodys Mnemonic
}
