% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_moodys.R
\name{process_moodys}
\alias{process_moodys}
\title{Process Moodys Data}
\usage{
process_moodys(df)
}
\arguments{
\item{df}{Data frame of Moody's data}
}
\value{
Data frame of Moody's data
}
\description{
Process Moody's data into long data frame. Calculate State and Denver Boulder MSA totals. Label county, state, msa and national and combine into single data frame.
}
