% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalks.R
\docType{data}
\name{region_xwalk}
\alias{region_xwalk}
\title{Region Crosswalk}
\format{
A data frame with 159 rows and 4 variables:
\describe{
\item{region_name}{Region Name}
\item{region_code}{Region Code}
\item{county_fips}{County FIPS code}
\item{county_name}{County name}
}
}
\source{
Colorado State Demography Office Crosswalk
}
\usage{
region_xwalk
}
\description{
A crosswalk file containing the counties in Colorado for mapping them to regions
}
\keyword{datasets}
