% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoecon-package.R
\docType{package}
\name{sdoecon-package}
\alias{sdoecon}
\alias{sdoecon-package}
\title{sdoecon: Package for loading SDO jobs data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gtottenSDO/sdoecon}
  \item Report bugs at \url{https://github.com/gtottenSDO/sdoecon/issues}
}

}
\author{
\strong{Maintainer}: Greg Totten \email{greg.totten@state.co.us}

}
\keyword{internal}
