% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_moodys_api_key.R
\name{set_moodys_api_key}
\alias{set_moodys_api_key}
\title{Install Moodys API Keys}
\usage{
set_moodys_api_key(accKey, encKey, permanent = TRUE)
}
\arguments{
\item{accKey}{Moodys API Access Key}

\item{encKey}{Moodys API Encryption Key}

\item{permanent}{Permanently write keys to user environment}
}
\description{
Install Moodys API Keys
}
