% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_moodys_series.R
\name{summarize_search_results}
\alias{summarize_search_results}
\title{Get Search Result Summary}
\usage{
summarize_search_results(search_df)
}
\arguments{
\item{search_df}{Data frame from convert_moodys_search()}
}
\value{
Prints summary of search results
}
\description{
Get Search Result Summary
}
